Issue/24 allow to download videos of chess960 and capablanca games

+++ src/EventListener/DownloadMp4Listener.php
@@ -24,7 +24,9 @@ class DownloadMp4Listener
         $currentRoute = $this->router->match($event->getRequest()->getPathInfo());
         if ('api_download_mp4' === $currentRoute['_route']) {
             $filename = $event->getRequest()->attributes->get('filename');
-            unlink(self::OUTPUT_FOLDER . '/'. $filename);
+            if (is_file(self::OUTPUT_FOLDER . '/'. $filename)) {
+                unlink(self::OUTPUT_FOLDER . '/'. $filename);
+            }
         }
     }
 }
